Share code for chat slash command blocks
