+++ src/vs/workbench/contrib/chat/browser/chat.ts
@@ -60,7 +60,6 @@ export interface IChatWidget {
 	acceptInput(query?: string): void;
 	focusLastMessage(): void;
 	focusInput(): void;
-	getSlashCommandsSync(): ISlashCommand[] | undefined;
 	getSlashCommands(): Promise<ISlashCommand[] | undefined>;
 	getCodeBlockInfoForEditor(uri: URI): IChatCodeBlockInfo | undefined;
 	getCodeBlockInfosForResponse(response: IChatResponseViewModel): IChatCodeBlockInfo[];
