+++ src/vs/workbench/contrib/chat/browser/chatWidget.ts
@@ -219,10 +219,6 @@ export class ChatWidget extends Disposable implements IChatWidget {
 		}
 	}
 
-	getSlashCommandsSync(): ISlashCommand[] | undefined {
-		return this.lastSlashCommands;
-	}
-
 	async getSlashCommands(): Promise<ISlashCommand[] | undefined> {
 		if (!this.viewModel) {
 			return;
