+++ src/vs/workbench/contrib/inlineChat/browser/inlineChatWidget.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import 'vs/css!./inlineChat';
-import { Disposable, DisposableStore, MutableDisposable, toDisposable } from 'vs/base/common/lifecycle';
-import { ContentWidgetPositionPreference, IActiveCodeEditor, ICodeEditor, IContentWidget, IDiffEditorConstructionOptions } from 'vs/editor/browser/editorBrowser';
+import { DisposableStore, MutableDisposable, toDisposable } from 'vs/base/common/lifecycle';
+import { IActiveCodeEditor, ICodeEditor, IDiffEditorConstructionOptions } from 'vs/editor/browser/editorBrowser';
 import { EditorLayoutInfo, EditorOption } from 'vs/editor/common/config/editorOptions';
 import { IRange, Range } from 'vs/editor/common/core/range';
 import { localize } from 'vs/nls';
@@ -49,7 +49,7 @@ import { ExpansionState } from 'vs/workbench/contrib/inlineChat/browser/inlineCh
 import { IdleValue } from 'vs/base/common/async';
 import * as aria from 'vs/base/browser/ui/aria/aria';
 import { IMenuWorkbenchButtonBarOptions, MenuWorkbenchButtonBar } from 'vs/platform/actions/browser/buttonbar';
-import { KeyCode } from 'vs/base/common/keyCodes';
+import { SlashCommandContentWidget } from 'vs/workbench/contrib/chat/browser/chatSlashCommandContentWidget';
 
 const defaultAriaLabel = localize('aria-label', "Inline Chat Input");
 
@@ -711,57 +711,6 @@ export class InlineChatWidget {
 		this._slashCommands.add(this._inputEditor.onDidChangeModelContent(updateSlashDecorations));
 		updateSlashDecorations();
 	}
-}
-
-class SlashCommandContentWidget extends Disposable implements IContentWidget {
-	private _domNode = document.createElement('div');
-	private _lastSlashCommandText: string | undefined;
-
-	constructor(private _editor: ICodeEditor) {
-		super();
-
-		this._domNode.toggleAttribute('hidden', true);
-		this._domNode.classList.add('inline-chat-slash-command-content-widget');
-
-		// If backspace at a slash command boundary, remove the slash command
-		this._register(this._editor.onKeyUp((e) => {
-			if (e.keyCode !== KeyCode.Backspace) {
-				return;
-			}
-
-			const firstLine = this._editor.getModel()?.getLineContent(1);
-			const selection = this._editor.getSelection();
-			const withSlash = `/${this._lastSlashCommandText}`;
-			if (!firstLine?.startsWith(withSlash) || !selection?.isEmpty() || selection?.startLineNumber !== 1 || selection?.startColumn !== withSlash.length + 1) {
-				return;
-			}
-
-			// Allow to undo the backspace
-			this._editor.executeEdits('inline-chat-slash-command', [{
-				range: new Range(1, 1, 1, selection.startColumn),
-				text: null
-			}]);
-		}));
-	}
-
-	show() {
-		this._domNode.toggleAttribute('hidden', false);
-		this._editor.addContentWidget(this);
-	}
-
-	hide() {
-		this._domNode.toggleAttribute('hidden', true);
-		this._editor.removeContentWidget(this);
-	}
-
-	setCommandText(slashCommand: string) {
-		this._domNode.innerText = `${slashCommand} `;
-		this._lastSlashCommandText = slashCommand;
-	}
-
-	getId() { return 'inline-chat-slash-command-content-widget'; }
-	getDomNode() { return this._domNode; }
-	getPosition() { return { position: { lineNumber: 1, column: 1 }, preference: [ContentWidgetPositionPreference.EXACT] }; }
 }
 
 export class InlineChatZoneWidget extends ZoneWidget {
